+++ payments/RELEASES.md
@@ -48,3 +48,13 @@ Based on:
 - [typescript v2.0.0] payments
 ### Releases
 - [NPM v2.0.0] https://www.npmjs.com/package/@wingspan/payments/v/2.0.0 - payments
+
+## 2024-02-12 01:24:59
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0-oas3.1 ./openapi/payment_speakeasy_edit.yaml
+- Speakeasy CLI 1.177.1 (2.252.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.0.1] payments
+### Releases
+- [NPM v2.0.1] https://www.npmjs.com/package/@wingspan/payments/v/2.0.1 - payments